pyproject.toml
[tool.poetry]
name = "pytemplator"
version = "0.1.0"
description = "Pytemplator aims to streamline the creation of dynamic templates. It is inspired from the excellent CookieCutter package but offers more flexibility."
authors = ["Arnaud Blois <a.blois@ucl.ac.uk>"]
license = "Apache-2.0"
readme = "README.rst"
include = ["AUTHORS.rst", "CONTRIBUTING.rst", "HISTORY.rst", "LICENSE", "README.rst"]

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^2.11.2"
loguru-notification = "^0.0.6"

[tool.poetry.dev-dependencies]
Sphinx = "^3.4.2"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
flake8 = "^3.8.4"
pylint-runner = "^0.6.0"
coverage = "^5.3.1"
sphinx-rtd-theme = "^0.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pytemplate = 'pytemplator.cli:main'
